refactor: split matching into own file
For sake of readability the css matching is moved to its own file
